perf(datacollection): refactor datacollection into mutiple classes
.... note that this commit does not yet include updated tests
