+++ ladybug/_datacollectionbase.py
@@ -1,0 +1,484 @@
+# coding=utf-8
+"""Base class for Ladybug Data Collections."""
+
+from __future__ import division
+
+from .header import Header
+from .dt import DateTime
+
+from string import ascii_lowercase
+import math
+import sys
+if (sys.version_info >= (3, 0)):
+    xrange = range
+
+
+class BaseCollection(object):
+    """Base class for all Data Collections."""
+
+    __slots__ = ('_header', '_values', '_datetimes')
+
+    def __init__(self, header, values, datetimes):
+        """Initialize base collection.
+
+        Args:
+            header: A Ladybug Header object.
+            values: A list of values.
+            datetimes: A list of Ladybug DateTime objects that aligns with
+                the list of values.
+        """
+        assert isinstance(header, Header), \
+            'header must be a Ladybug Header object. Got {}'.format(type(header))
+        assert isinstance(datetimes, (list, tuple)), \
+            'datetimes must be a list or tuple. Got {}'.format(type(datetimes))
+        if isinstance(datetimes, tuple):
+            datetimes = list(datetimes)
+
+        self._header = header
+        self._datetimes = datetimes
+        self.values = values
+
+    @classmethod
+    def from_json(cls, data):
+        """Create a Data Collection from a dictionary.
+
+        Args:
+            {
+                "header": A Ladybug Header,
+                "values": An array of values,
+                "datetimes": An array of datetimes
+            }
+        """
+        assert 'header' in data, 'Required keyword "header" is missing!'
+        assert 'values' in data, 'Required keyword "values" is missing!'
+        assert 'datetimes' in data, 'Required keyword "datetimes" is missing!'
+        return cls(Header.from_json(data['header']),
+                   data['values'],
+                   [DateTime.from_json(dat) for dat in data['datetimes']])
+
+    @property
+    def header(self):
+        """Return the header for this collection."""
+        return self._header
+
+    @property
+    def datetimes(self):
+        """Return datetimes for this collection as a tuple."""
+        return tuple(self._datetimes)
+
+    @property
+    def values(self):
+        """Return the Data Collection's list of numerical values."""
+        return tuple(self._values)
+
+    @values.setter
+    def values(self, values):
+        assert isinstance(values, (list, tuple)), \
+            'values must be a list or tuple. Got {}'.format(type(values))
+        if isinstance(values, tuple):
+            values = list(values)
+        assert len(values) == len(self.datetimes), \
+            'Length of values list must match length of datetimes list. {} != {}'.format(
+                len(values), len(self.datetimes))
+        self._values = values
+
+    @property
+    def bounds(self):
+        """Return a tuple as (min, max)."""
+        return (min(self._values), max(self._values))
+
+    @property
+    def min(self):
+        """Return the min of the Data Collection values."""
+        return min(self._values)
+
+    @property
+    def max(self):
+        """Return the max of the Data Collection values."""
+        return max(self._values)
+
+    @property
+    def average(self):
+        """Return the average of the Data Collection values."""
+        return sum(self._values) / len(self._values)
+
+    @property
+    def median(self):
+        """Return the median of the Data Collection values."""
+        return self._percentile(self._values, 50)
+
+    @property
+    def total(self):
+        """Return the total of the Data Collection values."""
+        return sum(self._values)
+
+    def duplicate(self):
+        """Return a copy of the current Data Collection."""
+        return self.__class__(
+            self.header.duplicate(), self.values, self.datetimes)
+
+    def get_aligned_collection(self, value=0, data_type=None, unit=None):
+        """Return a Collection aligned with this one composed of one repeated value.
+
+        Aligned Data Collections are of the same Data Collection class, have the same
+        number of values and have matching datetimes.
+
+        Args:
+            value: The value to be repeated in the aliged collection values.
+                Default: 0.
+            data_type: The data type of the aligned collection. Default is to
+                use the data type of this collection.
+            unit: The unit of the aligned collection. Default is to
+                use the unit of this collection.
+        """
+        data_type = data_type or self.header.data_type
+        unit = unit or self.header.unit
+        values = [value] * len(self._values)
+        header = Header(data_type, unit, self.header.analysis_period,
+                        self.header.metadata)
+        return self.__class__(header, values, self.datetimes)
+
+    def convert_to_unit(self, unit):
+        """Convert the Data Collection to the input unit."""
+        self._values = self._header.data_type.to_unit(
+            self._values, unit, self._header.unit)
+        self._header._unit = unit
+
+    def convert_to_ip(self):
+        """Convert the Data Collection to IP units."""
+        self._values, self._header._unit = self._header.data_type.to_ip(
+                self._values, self._header.unit)
+
+    def convert_to_si(self):
+        """Convert the Data Collection to SI units."""
+        self._values, self._header._unit = self._header.data_type.to_si(
+                self._values, self._header.unit)
+
+    def to_unit(self, unit):
+        """Return a Data Collection in the input unit."""
+        new_data_c = self.duplicate()
+        new_data_c.convert_to_unit(unit)
+        return new_data_c
+
+    def to_ip(self):
+        """Return a Data Collection in IP units."""
+        new_data_c = self.duplicate()
+        new_data_c.convert_to_ip()
+        return new_data_c
+
+    def to_si(self):
+        """Return a Data Collection in SI units."""
+        new_data_c = self.duplicate()
+        new_data_c.convert_to_si()
+        return new_data_c
+
+    def get_highest_values(self, count):
+        """Find the highest values in the Data Collection.
+
+        Args:
+            count: Integer representing the number of highest values to account for.
+
+        Returns:
+            highest_values: The n highest values in data list, ordered from
+                highest to lowest.
+            highest_values_index: Indicies of the n highest values in data
+                list, ordered from highest to lowest.
+        """
+        count = int(count)
+        assert count <= len(self._values), \
+            'count must be smaller than or equal to values length. {} > {}.'.format(
+                count, len(self._values))
+        assert count > 0, \
+            'count must be greater than 0. Got {}.'.format(count)
+        highest_values = sorted(self._values, reverse=True)[0:count]
+        highest_values_index = sorted(list(xrange(len(self._values))),
+                                      key=lambda k: self._values[k],
+                                      reverse=True)[0:count]
+        return highest_values, highest_values_index
+
+    def get_lowest_values(self, count):
+        """Find the lowest values in the Data Collection.
+
+        Args:
+            count: Integer representing the number of lowest values to account for.
+
+        Returns:
+            highest_values: The n lowest values in data list, ordered from
+                lowest to lowest.
+            lowest_values_index: Indicies of the n lowest values in data
+                list, ordered from lowest to lowest.
+        """
+        count = int(count)
+        assert count <= len(self._values), \
+            'count must be <= to Data Collection len. {} > {}.'.format(
+                count, len(self._values))
+        assert count > 0, \
+            'count must be greater than 0. Got {}.'.format(count)
+        lowest_values = sorted(self._values)[0:count]
+        lowest_values_index = sorted(list(xrange(len(self._values))),
+                                     key=lambda k: self._values[k])[0:count]
+        return lowest_values, lowest_values_index
+
+    def get_percentile(self, percentile):
+        """Filter the Data Collection based on a conditional statement.
+
+        Args:
+            percentile: A float value from 0 to 100 representing the
+                requested percentile.
+
+        Return:
+            The Data Collection value at the input percentile
+        """
+        return self._percentile(self._values, percentile)
+
+    def filter_by_conditional_statement(self, statement):
+        """Filter the Data Collection based on a conditional statement.
+
+        Args:
+            statement: A conditional statement as a string (e.g. a > 25 and a%5 == 0).
+                The variable should always be named as 'a' (without quotations).
+
+        Return:
+            A new Data Collection containing only the filtered data
+        """
+        _filt_values, _filt_datetimes = self._filter_by_statement(statement)
+        return self.__class__(self.header.duplicate(), _filt_values, _filt_datetimes)
+
+    def filter_by_pattern(self, pattern):
+        """Filter the Data Collection based on a list of booleans.
+
+        Args:
+            pattern: A list of True/False values.  Typically, this is a list
+                with a length matching the length of the Data Collections values
+                but it can also be a pattern to be repeated over the Data Collection.
+
+        Return:
+            A new Data Collection with filtered data
+        """
+        _filt_values, _filt_datetimes = self._filter_by_pattern(pattern)
+        return self.__class__(self.header.duplicate(), _filt_values, _filt_datetimes)
+
+    def is_collection_aligned(self, data_collection):
+        """Check if this Data Collection is aligned with another.
+
+        Aligned Data Collections are of the same Data Collection class, have the
+        same number of values and have matching datetimes.
+
+        Args:
+            data_collection: The Data Collection which you want to test if this
+                collection is aligned with.
+
+        Return:
+            True if collections are aligned, Fale if not aligned
+        """
+        if type(self) != type(data_collection):
+            return False
+        elif len(self.values) != len(data_collection.values):
+            return False
+        elif self.datetimes != data_collection.datetimes:
+            return False
+        else:
+            return True
+
+    @staticmethod
+    def filter_collections_by_statement(data_collections, statement):
+        """Generate a filtered data collections according to a conditional statement.
+
+        Args:
+            data_collections: A list of aligned Data Collections to be evaluated
+                against the statement.
+            statement: A conditional statement as a string (e.g. a>25 and a%5==0).
+                The variable should always be named as 'a' (without quotations).
+
+        Return:
+            collections: A list of Data Collections that have been filtered based
+                on the statement.
+        """
+        pattern = BaseCollection.pattern_from_collections_and_statement(
+            data_collections, statement)
+        collections = [coll.filter_by_pattern(pattern) for coll in data_collections]
+        return collections
+
+    @staticmethod
+    def pattern_from_collections_and_statement(data_collections, statement):
+        """Generate a list of booleans from data collections and a conditional statement.
+
+        Args:
+            data_collections: A list of aligned Data Collections to be evaluated
+                against the statement.
+            statement: A conditional statement as a string (e.g. a>25 and a%5==0).
+                The variable should always be named as 'a' (without quotations).
+
+        Return:
+            pattern: A list of True/False booleans with the length of the
+                Data Collections where True meets the conditional statement
+                and False does not.
+        """
+        BaseCollection.are_collections_aligned(data_collections)
+        correct_var = BaseCollection._check_conditional_statement(
+            statement, len(data_collections))
+
+        pattern = []
+        for i in xrange(len(data_collections[0])):
+            num_statement = statement
+            for j, coll in enumerate(data_collections):
+                var = correct_var[j]
+                num_statement = num_statement.replace(var, coll[i])
+            pattern.append(eval(num_statement, globals={}))
+        return pattern
+
+    @staticmethod
+    def are_collections_aligned(data_collections, raise_exception=True):
+        """Test if a series of Data Collections are aligned with one another.
+
+        Aligned Data Collections are of the same Data Collection class, have the
+        same number of values and have matching datetimes.
+
+        Args:
+            data_collections: A list of Data Collections for which you want to
+                test if they are al aligned with one another.
+
+        Return:
+            True if collections are aligned, Fale if not aligned
+        """
+        if len(data_collections) > 1:
+            first_coll = data_collections[0]
+            for coll in data_collections[1:]:
+                if not first_coll.is_collection_aligned(coll):
+                    if raise_exception is True:
+                        error_msg = '{} Data Collection is not aligned with '\
+                            '{} Data Collection.'.format(
+                                first_coll.header.data_type, coll.header.data_type)
+                        raise ValueError(error_msg)
+                    return False
+        return True
+
+    def _check_conditional_statement(self, statement, num_collections):
+        """Method to check conditional statements to be sure that they are valid.
+
+        Args:
+            statement: A conditional statement as a string (e.g. a>25 and a%5==0).
+                The variable should always be named as 'a' (without quotations).
+            num_collections: An integer representing the number of data collections
+                that the statement will be evaluating.
+
+        Return:
+            correct_var: A list of the correct variable names that should be
+                used within the statement (eg. ['a', 'b', 'c'])
+        """
+        # Determine what the list of variables should be based on the num_collections
+        correct_var = list(ascii_lowercase)[:num_collections]
+
+        # Clean out the oeprators of the statement
+        st_statement = statement.lower() \
+            .replace("and", "").replace("or", "") \
+            .replace("not", "").replace("in", "").replace("is", "")
+        parsed_st = [s for s in st_statement if s.isalpha()]
+
+        # Perform the chec
+        assert list(set(parsed_st)) == correct_var, \
+            'Invalid conditional statement: {}\n '\
+            'Statement should be a valid Python statement'\
+            ' and the variables should be named as follows: {}'.format(
+                statement, ', '.join(correct_var))
+        return correct_var
+
+    def _filter_by_statement(self, statement):
+        """Filter the data collection based on a conditional statement."""
+        self._check_conditional_statement(statement, 1)
+        statement = statement.replace('a', 'd')
+        _filt_values, _filt_datetimes = [], []
+        for i, d in enumerate(self._values):
+            if eval(statement, globals={'d': d}):
+                _filt_values.append(d)
+                _filt_datetimes.append(self.datetimes[i])
+        return _filt_values, _filt_datetimes
+
+    def _filter_by_pattern(self, pattern):
+        """Filter the Filter the Data Collection based on a list of booleans."""
+        try:
+            _len = len(pattern)
+        except TypeError:
+            raise TypeError("pattern is not a list of Booleans. Got {}".format(
+                type(pattern)))
+        _filt_values = [d for i, d in enumerate(self._values) if pattern[i % _len]]
+        _filt_datetimes = [d for i, d in enumerate(self.datetimes) if pattern[i % _len]]
+        return _filt_values, _filt_datetimes
+
+    def _is_in_data_type_range(self, raise_exception=True):
+        """Check if the Data Collection values are in permissable ranges for the data_type.
+
+        If this method returns False, the Data Collection's data is
+        physically or mathematically impossible for the data_type."""
+        return self._header.data_type.is_in_range(
+            self._values, self._header.unit, raise_exception)
+
+    def _is_in_epw_range(self, raise_exception=True):
+        """Check if Data Collection values are in permissable ranges for EPW files."""
+        return self._header.data_type.is_in_range_epw(
+            self._values, self._header.unit, raise_exception)
+
+    def _percentile(self, values, percent, key=lambda x: x):
+        """Find the percentile of a list of values.
+
+        Args:
+            values: A list of values for which percentiles are desired
+            percent: A float value from 0 to 100 representing the requested percentile.
+            key: optional key function to compute value from each element of N.
+
+        Return:
+            The percentile of the values
+        """
+        vals = sorted(values)
+        k = (len(vals) - 1) * (percent / 100)
+        f = math.floor(k)
+        c = math.ceil(k)
+        if f == c:
+            return key(vals[int(k)])
+        d0 = key(vals[int(f)]) * (c-k)
+        d1 = key(vals[int(c)]) * (k-f)
+        return d0+d1
+
+    def __len__(self):
+        return len(self._values)
+
+    def __getitem__(self, key):
+        return self._values[key]
+
+    def __setitem__(self, key, value):
+        self._values[key] = value
+
+    def __delitem__(self, key):
+        del self._values[key]
+        del self._datetimes[key]
+
+    def __iter__(self):
+        return iter(self._values)
+
+    def __reversed__(self):
+        return reversed(self._values)
+
+    def __contains__(self, item):
+        return item in self._values
+
+    def to_json(self):
+        """Convert Data Collection to a dictionary."""
+        return {
+            'header': self.header.to_json(),
+            'values': self._values,
+            'datetimes': [dat.to_json() for dat in self.datetimes]
+        }
+
+    @property
+    def isDataCollection(self):
+        """Return True."""
+        return True
+
+    def ToString(self):
+        """Overwrite .NET ToString method."""
+        return self.__repr__()
+
+    def __repr__(self):
+        """Discontinuous Collection representation."""
+        return "Discontinuous Data Collection\n{} ({})\n...{} values...".format(
+            self.header.data_type, self.header.unit, len(self._values))
